backend portafolioAF 1.09.04

+++ src/main/java/com/portfolioB/AF/Repository/IPersonaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPersonaRepository extends JpaRepository<Persona,Long> {
+public interface IPersonaRepository extends JpaRepository<Persona,Integer> {
     public Optional<Persona> findByName(String name);
     public boolean existsByName(String name);
